projects/time-tracking-system/engineering-full-stack-developer/code/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pyjwt==2.8.0
bcrypt==4.1.2
openpyxl==3.1.2
